#7622 Summary Restart : Add summary restart reader
Move libecl reader into RifEclEclipseSummary
Add a summary file restart history reader
Simplify RifEclipseSummary
