+++ ApplicationLibCode/FileInterface/CMakeLists_files.cmake
@@ -62,6 +62,8 @@ ${CMAKE_CURRENT_LIST_DIR}/RifStimPlanModelAsymmetricFrkExporter.h
 ${CMAKE_CURRENT_LIST_DIR}/RifSurfaceExporter.h
 ${CMAKE_CURRENT_LIST_DIR}/RifOpmCommonSummary.h
 ${CMAKE_CURRENT_LIST_DIR}/RifEnsembleFractureStatisticsExporter.h
+${CMAKE_CURRENT_LIST_DIR}/RifSummaryReaderMultipleFiles.h
+${CMAKE_CURRENT_LIST_DIR}/RifEclEclipseSummary.h
 
 # HDF5 file reader is directly included in ResInsight main CmakeList.txt
 #${CMAKE_CURRENT_LIST_DIR}/RifHdf5Reader.h
@@ -128,6 +130,8 @@ ${CMAKE_CURRENT_LIST_DIR}/RifStimPlanModelAsymmetricFrkExporter.cpp
 ${CMAKE_CURRENT_LIST_DIR}/RifSurfaceExporter.cpp
 ${CMAKE_CURRENT_LIST_DIR}/RifOpmCommonSummary.cpp
 ${CMAKE_CURRENT_LIST_DIR}/RifEnsembleFractureStatisticsExporter.cpp
+${CMAKE_CURRENT_LIST_DIR}/RifSummaryReaderMultipleFiles.cpp
+${CMAKE_CURRENT_LIST_DIR}/RifEclEclipseSummary.cpp
 
 # HDF5 file reader is directly included in ResInsight main CmakeList.txt
 #${CMAKE_CURRENT_LIST_DIR}/RifHdf5Reader.cpp
